avoid race condition - use final block hash
When sending a transaction, it should build on top of a final block.
Otherwise, we aer suspect to race conditions.

The RPC receiving the TX might already know the block hash,
while the chunk producer might still be waiting for approvals.

In test, it almost looks like even the same node may return a latest
hash that is newer than what was on the DB.
Not sure how that happens, but according to the timestamps in the logs,
it seems to be like that.