move state init struct into core-primitives
Moves  `DeterministicAccountStateInit` from near-primitives to
near-primitives-core. It also requires moving `GlobalContractIdentifier`
because it is a field of the former type.

The goal is to give the VM runtime access to this type and making it easier for
smart contracts to get access to the type, too.

While we don't want to clutter core-primitives more than necessary, this
is necessary to give the VM runtime access to this type. The VM runtime
wants to parse this struct on the host function to add state init actions
from a function call, mainly to know the number of entries.

One alternative would be to treat state init actions as opaque Vec<u8>.
Verification per action is done on the transaction / receipt runtime anyway.
But because gas costs depend on the number of entries in the BTreeMap,
the vm runtime still has to parse the struct .

Another alternative would be to specify `data: BTreeMap<Vec<u8>, Vec<u8>>`
though host functions. That makes for a much more complex host function
interface. Possible but I don't like it.

Finally, I realized that smart contracts need to construct this on-chain to verify
the deterministic predecessor_id against specific inputs. They will need the
full type. I think this should be shared code with what we use in nearcore,
rather than a re-implementation.

With all this in mind, I suggest to put the type in near-primitives-core.
Potentially, this could even  become a independent crate or be integrated
in near-account-id. But I will keep it in nearcore for now.